services/ai_engine/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
torch>=2.1.2
pandas>=2.2.0
numpy>=1.26.3
scikit-learn>=1.4.0
ccxt>=4.2.14
pandas_ta>=0.3.14b0
pydantic>=2.6.0
python-dotenv>=1.0.1
websockets>=12.0
httpx>=0.26.0
psycopg2-binary
sqlalchemy
asyncpg
ray[default]
google-cloud-aiplatform
google-cloud-bigquery
